godot path doesnt have to be there twice

+++ buildManually.ps1
@@ -1,5 +1,6 @@
 # Adjust the Godot path and arguments as per your local setup
 $godotPath = $env:GODOT
+$releaseDir = "Release"
 
 # Extract Version from Project File
 $version = Select-String -Path "Connect-A-PIC.csproj" -Pattern '<Version>(.*)</Version>' | % { $_.Matches.Groups[1].Value }
@@ -9,15 +10,13 @@ Write-Host "The extracted project version is $version"
 $godotVersion = Get-Content -Raw -Path 'global.json' | ConvertFrom-Json | Select-Object -ExpandProperty 'msbuild-sdks' | Select-Object -ExpandProperty 'Godot.NET.Sdk'
 Write-Host "Godot version: $godotVersion"
 
-# Adjust the Godot path and arguments as per your local setup
-$godotPath = $env:GODOT
-
 # Delete all old Releases
-Remove-Item -Path "Release\*" -Recurse -Force
-Write-Host "Release Folder got cleared"
+if (Test-Path $releaseDir) {
+    Remove-Item -Path "Release\*" -Recurse -Force
+    Write-Host "Release Folder got cleared"
+}
 
 # Create Build Directory
-$releaseDir = "Release"
 if (-not (Test-Path $releaseDir)) {
     New-Item -ItemType Directory -Path $releaseDir
 }
